close #153 - edit schema

+++ backend/entity/schema.go
@@ -197,7 +197,7 @@ type RecordPetition struct {
 type IncreaseGrades struct {
 	gorm.Model
 	Date        time.Time
-	grade_point float64
+	grade_point int16
 	Description string
 
 	StudentID *uint
